+++ tests/Stubs/ExampleInvalidPermissionFlags.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Reinder83\BinaryFlags\Tests\Stubs;
+
+use Reinder83\BinaryFlags\BinaryEnumFlags;
+
+/**
+ * @extends BinaryEnumFlags<InvalidPermission>
+ */
+class ExampleInvalidPermissionFlags extends BinaryEnumFlags
+{
+    protected static function getFlagEnumClass(): string
+    {
+        return InvalidPermission::class;
+    }
+}